requirements.txt
Pillow==10.1.0
Pillow==10.2.0
PyAutoGUI==0.9.54
pynput==1.7.6
